Utilize CURL_WANTS_CA_BUNDLE_ENV in the curl build configuration.
This causes libcurl to get the CA cert bundle location from the
CA_BUNDLE environment variable, as opposed to it being hardcoded at
compile time.

Enabling this allows for flexibility when the Python binary packages are
deployed to various Linux distributions, which place the CA cert bundle
in different locations.